@@ -14,8 +14,8 @@
 package com.querydsl.sql.spring;
 
 import java.sql.Connection;
+import java.util.function.Supplier;
 
-import javax.inject.Provider;
 import javax.sql.DataSource;
 
 import org.springframework.jdbc.datasource.DataSourceUtils;
@@ -31,7 +31,7 @@ import org.springframework.jdbc.datasource.DataSourceUtils;
  * }
  * </pre>
  */
-public class SpringConnectionProvider implements Provider<Connection> {
+public class SpringConnectionProvider implements Supplier<Connection> {
 
     private final DataSource dataSource;
 
