@@ -25,9 +25,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import java.util.function.Supplier;
 
 import org.jetbrains.annotations.Nullable;
-import javax.inject.Provider;
 
 import com.mysema.commons.lang.CloseableIterator;
 import com.querydsl.core.*;
@@ -55,7 +55,7 @@ public abstract class AbstractSQLQuery<T, Q extends AbstractSQLQuery<T, Q>> exte
     private static final QueryFlag rowCountFlag = new QueryFlag(QueryFlag.Position.AFTER_PROJECTION, ", count(*) over() ");
 
     @Nullable
-    private Provider<Connection> connProvider;
+    private Supplier<Connection> connProvider;
 
     @Nullable
     private Connection conn;
@@ -83,11 +83,11 @@ public abstract class AbstractSQLQuery<T, Q extends AbstractSQLQuery<T, Q>> exte
         this.useLiterals = configuration.getUseLiterals();
     }
 
-    public AbstractSQLQuery(Provider<Connection> connProvider, Configuration configuration) {
+    public AbstractSQLQuery(Supplier<Connection> connProvider, Configuration configuration) {
         this(connProvider, configuration, new DefaultQueryMetadata());
     }
 
-    public AbstractSQLQuery(Provider<Connection> connProvider, Configuration configuration, QueryMetadata metadata) {
+    public AbstractSQLQuery(Supplier<Connection> connProvider, Configuration configuration, QueryMetadata metadata) {
         super(new QueryMixin<Q>(metadata, false), configuration);
         this.connProvider = connProvider;
         this.listeners = new SQLListeners(configuration.getListeners());
