@@ -20,9 +20,9 @@ import java.util.List;
 import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import java.util.function.Supplier;
 
 import org.jetbrains.annotations.Nullable;
-import javax.inject.Provider;
 
 import com.querydsl.core.QueryMetadata;
 import com.querydsl.core.dml.DMLClause;
@@ -49,7 +49,7 @@ public abstract class AbstractSQLClause<C extends AbstractSQLClause<C>> implemen
     protected SQLListenerContextImpl context;
 
     @Nullable
-    private Provider<Connection> connProvider;
+    private Supplier<Connection> connProvider;
 
     @Nullable
     private Connection conn;
@@ -60,7 +60,7 @@ public abstract class AbstractSQLClause<C extends AbstractSQLClause<C>> implemen
         this.useLiterals = configuration.getUseLiterals();
     }
 
-    public AbstractSQLClause(Configuration configuration, Provider<Connection> connProvider) {
+    public AbstractSQLClause(Configuration configuration, Supplier<Connection> connProvider) {
         this(configuration);
         this.connProvider = connProvider;
     }
