@@ -18,9 +18,9 @@ import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.*;
 import java.util.logging.Logger;
+import java.util.function.Supplier;
 
 import org.jetbrains.annotations.Range;
-import javax.inject.Provider;
 
 import com.querydsl.core.*;
 import com.querydsl.core.QueryFlag.Position;
@@ -65,7 +65,7 @@ public abstract class AbstractSQLDeleteClause<C extends AbstractSQLDeleteClause<
         metadata.setValidatingVisitor(validatingVisitor);
     }
 
-    public AbstractSQLDeleteClause(Provider<Connection> connection, Configuration configuration, RelationalPath<?> entity) {
+    public AbstractSQLDeleteClause(Supplier<Connection> connection, Configuration configuration, RelationalPath<?> entity) {
         super(configuration, connection);
         this.entity = entity;
         metadata.addJoin(JoinType.DEFAULT, entity);
