@@ -14,8 +14,7 @@
 package com.querydsl.sql.dml;
 
 import java.sql.Connection;
-
-import javax.inject.Provider;
+import java.util.function.Supplier;
 
 import com.querydsl.sql.Configuration;
 import com.querydsl.sql.RelationalPath;
@@ -46,11 +45,11 @@ public class SQLInsertClause extends AbstractSQLInsertClause<SQLInsertClause> {
         super(connection, configuration, entity);
     }
 
-    public SQLInsertClause(Provider<Connection> connection, Configuration configuration, RelationalPath<?> entity, SQLQuery<?> subQuery) {
+    public SQLInsertClause(Supplier<Connection> connection, Configuration configuration, RelationalPath<?> entity, SQLQuery<?> subQuery) {
         super(connection, configuration, entity, subQuery);
     }
 
-    public SQLInsertClause(Provider<Connection> connection, Configuration configuration, RelationalPath<?> entity) {
+    public SQLInsertClause(Supplier<Connection> connection, Configuration configuration, RelationalPath<?> entity) {
         super(connection, configuration, entity);
     }
 }
