@@ -14,8 +14,7 @@
 package com.querydsl.sql.mssql;
 
 import java.sql.Connection;
-
-import javax.inject.Provider;
+import java.util.function.Supplier;
 
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.Expression;
@@ -30,15 +29,15 @@ import com.querydsl.sql.*;
  */
 public class SQLServerQueryFactory extends AbstractSQLQueryFactory<SQLServerQuery<?>> {
 
-    public SQLServerQueryFactory(Configuration configuration, Provider<Connection> connection) {
+    public SQLServerQueryFactory(Configuration configuration, Supplier<Connection> connection) {
         super(configuration, connection);
     }
 
-    public SQLServerQueryFactory(Provider<Connection> connection) {
+    public SQLServerQueryFactory(Supplier<Connection> connection) {
         this(new Configuration(new SQLServerTemplates()), connection);
     }
 
-    public SQLServerQueryFactory(SQLTemplates templates, Provider<Connection> connection) {
+    public SQLServerQueryFactory(SQLTemplates templates, Supplier<Connection> connection) {
         this(new Configuration(templates), connection);
     }
 
