@@ -14,8 +14,7 @@
 package com.querydsl.sql.postgresql;
 
 import java.sql.Connection;
-
-import javax.inject.Provider;
+import java.util.function.Supplier;
 
 import com.querydsl.core.DefaultQueryMetadata;
 import com.querydsl.core.QueryMetadata;
@@ -50,11 +49,11 @@ public class PostgreSQLQuery<T> extends AbstractPostgreSQLQuery<T, PostgreSQLQue
         super(conn, configuration, metadata);
     }
 
-    public PostgreSQLQuery(Provider<Connection> connProvider, Configuration configuration, QueryMetadata metadata) {
+    public PostgreSQLQuery(Supplier<Connection> connProvider, Configuration configuration, QueryMetadata metadata) {
         super(connProvider, configuration, metadata);
     }
 
-    public PostgreSQLQuery(Provider<Connection> connProvider, Configuration configuration) {
+    public PostgreSQLQuery(Supplier<Connection> connProvider, Configuration configuration) {
         super(connProvider, configuration, new DefaultQueryMetadata());
     }
 
