@@ -14,8 +14,7 @@
 package com.querydsl.sql.teradata;
 
 import java.sql.Connection;
-
-import javax.inject.Provider;
+import java.util.function.Supplier;
 
 import com.querydsl.core.DefaultQueryMetadata;
 import com.querydsl.core.QueryFlag;
@@ -56,11 +55,11 @@ public class TeradataQuery<T> extends AbstractTeradataQuery<T, TeradataQuery<T>>
         super(conn, configuration, metadata);
     }
 
-    public TeradataQuery(Provider<Connection> connProvider, Configuration configuration, QueryMetadata metadata) {
+    public TeradataQuery(Supplier<Connection> connProvider, Configuration configuration, QueryMetadata metadata) {
         super(connProvider, configuration, metadata);
     }
 
-    public TeradataQuery(Provider<Connection> connProvider, Configuration configuration) {
+    public TeradataQuery(Supplier<Connection> connProvider, Configuration configuration) {
         super(connProvider, configuration);
     }
 
