@@ -14,8 +14,7 @@
 package com.querydsl.sql.teradata;
 
 import java.sql.Connection;
-
-import javax.inject.Provider;
+import java.util.function.Supplier;
 
 import com.querydsl.core.Tuple;
 import com.querydsl.core.types.Expression;
@@ -30,15 +29,15 @@ import com.querydsl.sql.*;
  */
 public class TeradataQueryFactory extends AbstractSQLQueryFactory<TeradataQuery<?>> {
 
-    public TeradataQueryFactory(Configuration configuration, Provider<Connection> connection) {
+    public TeradataQueryFactory(Configuration configuration, Supplier<Connection> connection) {
         super(configuration, connection);
     }
 
-    public TeradataQueryFactory(Provider<Connection> connection) {
+    public TeradataQueryFactory(Supplier<Connection> connection) {
         this(new Configuration(new TeradataTemplates()), connection);
     }
 
-    public TeradataQueryFactory(SQLTemplates templates, Provider<Connection> connection) {
+    public TeradataQueryFactory(SQLTemplates templates, Supplier<Connection> connection) {
         this(new Configuration(templates), connection);
     }
 
