@@ -17,8 +17,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 import java.sql.Connection;
-
-import javax.inject.Provider;
+import java.util.function.Supplier;
 
 import org.easymock.EasyMock;
 import org.junit.Before;
@@ -35,12 +34,7 @@ public class MySQLQueryFactoryTest {
 
     @Before
     public void setUp() {
-        Provider<Connection> provider = new Provider<Connection>() {
-            @Override
-            public Connection get() {
-                return EasyMock.createNiceMock(Connection.class);
-            }
-        };
+        Supplier<Connection> provider = () -> EasyMock.createNiceMock(Connection.class);
         queryFactory = new MySQLQueryFactory(SQLTemplates.DEFAULT, provider);
     }
 
